Cerrar el popup al enviar el formulario de crear y editar puntos

+++ webapp/src/components/map/DeletePoint.js
@@ -1,7 +1,8 @@
 import {deletePoints} from '../../helper/PodHelper'
 
 export default function DeletePoint(pointId, marker, map, session, webId){
-    deletePoints(session, webId, pointId).then(
-        map.removeLayer(marker)
-    );
+    if (window.confirm("Estás seguro de que deseas borrar el punto?"))
+        deletePoints(session, webId, pointId).then(
+            map.removeLayer(marker)
+        );
 }
